# backend/openapi.json
{
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "title": "Detail",
                        "type": "array"
                    }
                },
                "title": "HTTPValidationError",
                "type": "object"
            },
...
            "NotebookEntryIn": {
                "properties": {
                    "content": {
                        "maxLength": 2000,
                        "minLength": 1,
                        "title": "Content",
                        "type": "string"
                    },
                    "course_slug": {
                        "default": "voa-level-1",
                        "maxLength": 100,
                        "minLength": 1,
                        "title": "Course Slug",
                        "type": "string"
                    },
                    "kind": {
                        "enum": [
                            "note",
                            "favorite_phrase",
                            "personal_example",
                            "recurring_error",
                            "teacher_question"
                        ],
                        "title": "Kind",
                        "type": "string"
                    },
                    "lesson_number": {
                        "minimum": 1.0,
                        "title": "Lesson Number",
                        "type": "integer"
                    }
                },
                "required": [
                    "lesson_number",
                    "kind",
                    "content"
                ],
                "title": "NotebookEntryIn",
                "type": "object"
            },
            "NotebookEntryOut": {
                "properties": {
                    "content": {
                        "title": "Content",
                        "type": "string"
                    },
                    "course_slug": {
                        "title": "Course Slug",
                        "type": "string"
                    },
                    "course_title": {
                        "title": "Course Title",
                        "type": "string"
                    },
                    "created_at": {
                        "format": "date-time",
                        "title": "Created At",
                        "type": "string"
                    },
                    "id": {
                        "title": "Id",
                        "type": "integer"
                    },
                    "kind": {
                        "enum": [
                            "note",
                            "favorite_phrase",
                            "personal_example",
                            "recurring_error",
                            "teacher_question"
                        ],
                        "title": "Kind",
                        "type": "string"
                    },
                    "lesson_number": {
                        "title": "Lesson Number",
                        "type": "integer"
                    },
                    "lesson_title": {
                        "title": "Lesson Title",
                        "type": "string"
                    },
                    "unit_slug": {
                        "title": "Unit Slug",
                        "type": "string"
                    },
                    "updated_at": {
                        "format": "date-time",
                        "title": "Updated At",
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "course_slug",
                    "course_title",
                    "unit_slug",
                    "lesson_number",
                    "lesson_title",
                    "kind",
                    "content",
                    "created_at",
                    "updated_at"
                ],
                "title": "NotebookEntryOut",
                "type": "object"
            },
            "NotebookEntryUpdate": {
                "properties": {
                    "content": {
                        "maxLength": 2000,
                        "minLength": 1,
                        "title": "Content",
                        "type": "string"
                    },
                    "kind": {
                        "enum": [
                            "note",
                            "favorite_phrase",
                            "personal_example",
                            "recurring_error",
                            "teacher_question"
                        ],
                        "title": "Kind",
                        "type": "string"
                    }
                },
                "required": [
                    "kind",
                    "content"
                ],
                "title": "NotebookEntryUpdate",
                "type": "object"
            },
...
            "ValidationError": {
                "properties": {
                    "ctx": {
                        "title": "Context",
                        "type": "object"
                    },
                    "input": {
                        "title": "Input"
                    },
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "title": "Location",
                        "type": "array"
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                },
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError",
                "type": "object"
            },
...
        }
    },
    "info": {
        "title": "Aulas de Inglês — API",
        "version": "0.1.0"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/api/me/notebook": {
            "get": {
                "operationId": "listar_caderno_api_me_notebook_get",
                "parameters": [
                    {
                        "in": "query",
                        "name": "lesson",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "minimum": 1,
                                    "type": "integer"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Lesson"
                        }
                    },
                    {
                        "in": "query",
                        "name": "kind",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "enum": [
                                        "note",
                                        "favorite_phrase",
                                        "personal_example",
                                        "recurring_error",
                                        "teacher_question"
                                    ],
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Kind"
                        }
                    },
                    {
                        "in": "query",
                        "name": "course",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "maxLength": 100,
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Course"
                        }
                    },
                    {
                        "in": "query",
                        "name": "unit",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "maxLength": 100,
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Unit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/NotebookEntryOut"
                                    },
                                    "title": "Response Listar Caderno Api Me Notebook Get",
                                    "type": "array"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Listar Caderno",
                "tags": [
                    "me"
                ]
            },
            "post": {
                "operationId": "criar_anotacao_api_me_notebook_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookEntryIn"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookEntryOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Criar Anotacao",
                "tags": [
                    "me"
                ]
            }
        },
        "/api/me/notebook/{entry_id}": {
            "delete": {
                "operationId": "excluir_anotacao_api_me_notebook__entry_id__delete",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "required": true,
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Excluir Anotacao",
                "tags": [
                    "me"
                ]
            },
            "put": {
                "operationId": "atualizar_anotacao_api_me_notebook__entry_id__put",
                "parameters": [
                    {
                        "in": "path",
                        "name": "entry_id",
                        "required": true,
                        "schema": {
                            "title": "Entry Id",
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotebookEntryUpdate"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotebookEntryOut"
                                }
                            }
                        },
                        "description": "Successful Response"
                    },
                    "422": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        },
                        "description": "Validation Error"
                    }
                },
                "summary": "Atualizar Anotacao",
                "tags": [
                    "me"
                ]
            }
        },
...
    }
}